requirements.txt
# Core MCP dependencies
mcp==1.2.0
fastmcp
python-dotenv==1.0.1

# Gmail API dependencies
google-auth>=2.34.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.147.0

# Client dependencies (for testing with Anthropic API)
anthropic==0.42.0

# Server dependencies
uvicorn==0.34.0
starlette==0.45.2
sse-starlette==2.2.1

# Supporting libraries
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.8.2
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0